GitHub issue#1462, Code refactor to load cleaner JSON files instead of repeated JS functions + Moved function to find title inside index.js + Updated imports in all json configs
