+++ src/config/projectQuestions/computer_vision.v1.0.js
@@ -1,18 +1,11 @@
 import _ from 'lodash'
 // import NumberText from '../../components/NumberText/NumberText'
-import { findProduct, isFileRequired } from '../projectWizard'
+import { isFileRequired, findTitle } from '../projectWizard'
 
 const sections = [
   {
     id: 'appDefinition',
-    title: (project, showProduct) => {
-      const product = _.get(project, 'details.products[0]')
-      if (showProduct && product) {
-        const prd = findProduct(product)
-        if (prd) return prd.name
-      }
-      return 'Definition'
-    },
+    title: findTitle,
     required: true,
     description: 'Please answer a few basic questions about your project. You can also provide the needed information in a supporting document--add a link in the notes section or upload it below.',
     subSections: [
