+++ src/config/projectQuestions/real_world_testing.v1.0.js
@@ -1,7 +1,7 @@
 import React from 'react' // eslint-disable-line no-unused-vars
 import _ from 'lodash'
 import { Icons } from 'appirio-tech-react-components'
-import { findProduct, isFileRequired } from '../projectWizard'
+import { isFileRequired, findTitle } from '../projectWizard'
 
 import IconDontKnow from '../../assets/icons/icon-dont-know.svg'
 import IconTestStructured from '../../assets/icons/icon-test-structured.svg'
@@ -10,14 +10,7 @@ import IconTestUnstructured from '../../assets/icons/icon-test-unstructured.svg'
 const sections = [
   {
     id: 'appDefinition',
-    title: (project, showProduct) => {
-      const product = _.get(project, 'details.products[0]')
-      if (showProduct && product) {
-        const prd = findProduct(product)
-        if (prd) return prd.name
-      }
-      return 'Definition'
-    },
+    title: findTitle,
     required: true,
     description: 'Please answer a few basic questions about your project. You can also provide the needed information in a supporting document--add a link in the notes section or upload it below.',
     subSections: [
